backend/requirements.txt
asgiref==3.5.2

cffi==1.15.1
Django==3.2
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-httpsignature==1.0.0
djangorestframework-simplejwt==5.2.2
drf-orjson-renderer==1.7.1
gunicorn==20.1.0
httpsig==1.3.0
orjson==3.8.3
Pillow==9.3.0
psycopg2-binary==2.9.5
pycparser==2.21
pycryptodome==3.17
PyJWT==2.6.0
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
django-extensions==3.2.1